Made "custom" explosions on by default
Custom explosions would now be on by default in worlds - I can't really think of a reason for the majority of people to only want explosions not made by plugins to heal, if someone does want that they can toggle it. For those who don't want that, it's a potential problem to have it so that they forgot to enable it in some world, and some stuff that exploded ends up being exploded for good.

An alternative to this could be having the ability to change the default value via config, and also being able to change a value for ALL worlds via command?