Update to 1.16
Use the new Materials and the new BlockData in a lot of places. I didn't
check that everything works, and I know that rails don't come back like
they were before, or that glass panes and double chests don't reattach
perfectly. I didn't update the properties of many new blocks (lantern
has to be placed after the blocks above and below, soul torch is like a
torch and so on). But overall, it should do 95% of the job right, which
is still better than no plugin.